requirements.txt
phidata
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
googlesearch-python 
pycountry